+++ src/api/cloud/project/ai/capabilities/cloud-project-ai-capabilities.service.js
@@ -1,0 +1,5 @@
+angular
+  .module('ovh-api-services')
+  .service('OvhApiCloudProjectAiCapabilities', ($injector) => ({
+    Serving: () => $injector.get('OvhApiCloudProjectAiCapabilitiesServing'),
+  }));