+++ src/api/cloud/project/ai/capabilities/serving/flavor/cloud-project-ai-capabilities-serving-flavor.service.js
@@ -1,0 +1,5 @@
+angular
+  .module('ovh-api-services')
+  .service('OvhApiCloudProjectAiCapabilitiesServingFlavor', ($injector) => ({
+    v6: () => $injector.get('OvhApiCloudProjectAiCapabilitiesServingFlavorV6'),
+  }));