+++ src/api/cloud/project/ai/capabilities/serving/presetImage/cloud-project-ai-capabilities-serving-preset-image.v6.service.js
@@ -1,0 +1,9 @@
+angular
+  .module('ovh-api-services')
+  .service('OvhApiCloudProjectAiCapabilitiesServingPresetImageV6', ($resource) => {
+    const resource = $resource('/cloud/project/:serviceName/ai/capabilities/serving/presetImage', {
+      serviceName: '@serviceName',
+    });
+
+    return resource;
+  });