+++ src/api/email/exchange/service/server/email-exchange-service-server.v6.service.js
@@ -1,0 +1,12 @@
+angular.module('ovh-api-services')
+  .service('OvhApiEmailExchangeServiceServerV6', ($resource) => $resource(
+    '/email/exchange/:organizationName/service/:exchangeService/server',
+    {
+      organizationName: '@organizationName',
+      exchangeService: '@exchangeService',
+    },
+    {
+      get: { method: 'GET' },
+      update: { method: 'PUT' },
+    },
+  ));