+++ src/api/hosting/web/database/database.v6.service.js
@@ -1,6 +1,4 @@
-angular.module('ovh-api-services').service('OvhApiHostingWebDatabaseV6', ($resource) => {
-  return $resource('/hosting/web/:serviceName/database/:name', {
-    serviceName: '@serviceName',
-    name: '@name',
-  });
-});
+angular.module('ovh-api-services').service('OvhApiHostingWebDatabaseV6', ($resource) => $resource('/hosting/web/:serviceName/database/:name', {
+  serviceName: '@serviceName',
+  name: '@name',
+}));