+++ src/api/services/detach/services-detach.v6.service.js
@@ -1,0 +1,19 @@
+angular
+  .module('ovh-api-services')
+  .service('OvhApiServicesDetachV6', ($resource) => $resource(
+    '/services/:serviceId/detach/:planCode',
+    {
+      planCode: '@planCode',
+      serviceId: '@serviceId',
+    },
+    {
+      execute: {
+        url: '/services/:serviceId/detach/:planCode/execute',
+        method: 'POST',
+      },
+      simulate: {
+        url: '/services/:serviceId/detach/:planCode/simulate',
+        method: 'POST',
+      },
+    },
+  ));