+++ src/api/xdsl/modem/xdsl-modem.v6.service.js
@@ -27,5 +27,9 @@ angular.module('ovh-api-services').service('OvhApiXdslModemV6', ($resource, OvhA
       url: '/xdsl/:xdslId/modem/comfortExchange',
       interceptor,
     },
+    getComfortExchange: {
+      method: 'GET',
+      url: '/xdsl/:xdslId/modem/comfortExchange',
+    },
   });
 });